// m3-foundation/types/components/surface.d.ts
export type Length = string | number

export type Mode = 'standard' | 'modal'

export type Variant =
  | 'auto'
  | 'surface'
  | 'surface-dim'
  | 'surface-bright'
  | 'surface-container-lowest'
  | 'surface-container-low'
  | 'surface-container'
  | 'surface-container-high'
  | 'surface-container-highest'

export type Anchor =
  | 'none'
  | 'center'
  | 'start'
  | 'end'
  | 'top'
  | 'bottom'
  | 'top-start'
  | 'top-end'
  | 'bottom-start'
  | 'bottom-end'